chore: split projects service
