+++ docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.md
@@ -75,6 +75,12 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - infra/cloudflare/src/core/ai-rewrite/prompts.js
 - infra/cloudflare/src/core/ai-rewrite/testing.js
 - infra/cloudflare/src/core/ai-rewrite/text.js
+- infra/cloudflare/src/service/projects.js
+- infra/cloudflare/src/service/projects/airtable.js
+- infra/cloudflare/src/service/projects/auth.js
+- infra/cloudflare/src/service/projects/csv.js
+- infra/cloudflare/src/service/projects/github-csv.js
+- infra/cloudflare/src/service/projects/normalisation.js
 - public/components/guides/api.js
 - public/components/guides/front-matter.js
 - public/components/guides/guides-page.js
@@ -83,6 +89,7 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - tests/govuk-pages-render-workflow-state.test.js
 - tests/ai-rewrite-split-route-state.test.js
 - tests/mural-service-split-route-state.test.js
+- tests/projects-service-split-route-state.test.js
 - tests/study-guides-route-state.test.js
 - docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.md
 - docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.json
@@ -97,6 +104,7 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - Follow-up Lychee remediation corrected stale documentation links exposed by the narrower Git ignore rules, fixed the project notes tab route, and added narrow Lychee exclusions for placeholder-heavy Mural example transcripts and the Mustache HTML head partial.
 - Follow-up Mural service split kept `MuralServicePart` as the public route surface while moving board registry resolution, direct journal sticky writing, token refresh, viewer URL probing, and workspace/room setup into responsibility-specific modules.
 - Follow-up AI rewrite split kept `ai-rewrite.js` as the public route and Worker entry surface while moving defaults, HTTP helpers, text sanitisation, guardrails, prompt content, and testing helpers into responsibility-specific modules under `infra/cloudflare/src/core/ai-rewrite/`.
+- Follow-up projects service split kept `projects.js` as the public route operation layer while moving CSV parsing, GitHub CSV fallback, field normalisation/mapping, auth visibility checks, and Airtable access helpers into responsibility-specific modules under `infra/cloudflare/src/service/projects/`.
 
 ## Validation attempted
 
@@ -121,6 +129,13 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - Follow-up AI rewrite split: `npm run lint` — passed with existing warnings only.
 - Follow-up AI rewrite split: `npm run format:check` — passed after generated CSS was regenerated by lint.
 - Follow-up AI rewrite split: `npm run trace:coverage` — passed.
+- Follow-up projects service split: `node --check infra/cloudflare/src/service/projects.js && node --check infra/cloudflare/src/service/projects/airtable.js && node --check infra/cloudflare/src/service/projects/auth.js && node --check infra/cloudflare/src/service/projects/csv.js && node --check infra/cloudflare/src/service/projects/github-csv.js && node --check infra/cloudflare/src/service/projects/normalisation.js` — passed.
+- Follow-up projects service split: `node --test tests/projects-service-split-route-state.test.js tests/projects-parse-csv.test.js` — passed.
+- Follow-up projects service split: dynamic import of `infra/cloudflare/src/service/projects.js` confirmed existing public exports remain available — passed.
+- Follow-up projects service split: `npm test` — passed, 181 tests.
+- Follow-up projects service split: `npm run lint` — passed with existing warnings only.
+- Follow-up projects service split: `npm run format:check` — passed.
+- Follow-up projects service split: `npm run trace:coverage` — passed.
 
 ## Residual risk
